+++ cli_options/CHANGELOG.md
@@ -8,8 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-* If an option is declared as `:hidden` it will not be included in the docs.
-* Support deprecating options.
+* If an option is set with `doc: false` it is not included in docs.
+* Support deprecating options with `:deprecated`.
 
 ## [v0.1.0](https://github.com/sportradar/elixir-workspace/tree/cli_options/v0.1.0) (2024-05-13)